# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateRetailProductRequest" : {
        "properties" : {
          "category_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "code" : {
            "description" : "Unique ignoring case and surrounding spaces",
            "maxLength" : 40,
            "minLength" : 0,
            "type" : "string"
          },
          "cost_minor" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          },
          "description" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "sell_minor" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          },
          "unit_id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "category_id", "code", "cost_minor", "description", "sell_minor", "unit_id" ]
      },
...
      "RetailCategory" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "RetailCategoryList" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/RetailCategory"
            },
            "type" : "array"
          }
        }
      },
      "RetailCategoryRequest" : {
        "properties" : {
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
...
      "RetailPriceChange" : {
        "properties" : {
          "at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "by" : {
            "format" : "uuid",
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "new_cost_minor" : {
            "description" : "Present only with retail.profit.read",
            "format" : "int64",
            "type" : "integer"
          },
          "new_sell_minor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "old_cost_minor" : {
            "description" : "Present only with retail.profit.read",
            "format" : "int64",
            "type" : "integer"
          },
          "old_sell_minor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reason" : {
            "type" : "string"
          },
          "source" : {
            "description" : "initial, manual, purchase or import",
            "type" : "string"
          },
          "source_id" : {
            "format" : "uuid",
            "type" : "string"
          }
        }
      },
      "RetailPriceEditRequest" : {
        "description" : "At least one price; FR-RET-02",
        "properties" : {
          "cost_minor" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          },
          "reason" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "sell_minor" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        },
        "required" : [ "reason" ]
      },
      "RetailPriceHistory" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/RetailPriceChange"
            },
            "type" : "array"
          }
        }
      },
      "RetailProduct" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "category" : {
            "type" : "string"
          },
          "category_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "cost_minor" : {
            "description" : "Present only with retail.profit.read",
            "format" : "int64",
            "type" : "integer"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "negative" : {
            "description" : "True when the branch balance is below zero",
            "type" : "boolean"
          },
          "qty" : {
            "description" : "The branch balance, when branch_id was given",
            "type" : "string"
          },
          "sell_minor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "unit" : {
            "type" : "string"
          },
          "unit_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
      "RetailProductPage" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/RetailProduct"
            },
            "type" : "array"
          },
          "next_cursor" : {
            "type" : "string"
          }
        }
      },
...
      "RetailUnit" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "RetailUnitList" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/RetailUnit"
            },
            "type" : "array"
          }
        }
      },
      "RetailUnitRequest" : {
        "properties" : {
          "name" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
...
      "TenantSettings" : {
        "properties" : {
          "allow_loans_before_kyc_verified" : {
            "type" : "boolean"
          },
          "appraisal_weights" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "object"
          },
          "approval_thresholds_minor" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "approval_validity_days" : {
            "format" : "int32",
            "type" : "integer"
          },
          "disabled_collateral_types" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "display_name" : {
            "type" : "string"
          },
          "max_active_loans_per_member" : {
            "format" : "int32",
            "type" : "integer"
          },
          "receipt_footer" : {
            "type" : "string"
          },
          "require_mfa_all_staff" : {
            "type" : "boolean"
          },
          "retail_allow_negative_stock" : {
            "type" : "boolean"
          },
          "sms_sender_name" : {
            "type" : "string"
          },
          "sms_window_end" : {
            "type" : "string"
          },
          "sms_window_start" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "UpdateRetailProductRequest" : {
        "description" : "Omitted fields are unchanged; prices change only through /prices",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "category_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "code" : {
            "maxLength" : 40,
            "minLength" : 1,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 300,
            "minLength" : 1,
            "type" : "string"
          },
          "unit_id" : {
            "format" : "uuid",
            "type" : "string"
          }
        }
      },
      "UpdateTenantSettingsRequest" : {
        "properties" : {
          "allow_loans_before_kyc_verified" : {
            "type" : "boolean"
          },
          "appraisal_weights" : {
            "additionalProperties" : {
              "description" : "repayment_history, affordability, collateral_cover, exposure; summing to 100",
              "format" : "int32",
              "type" : "integer"
            },
            "description" : "repayment_history, affordability, collateral_cover, exposure; summing to 100",
            "type" : "object"
          },
          "approval_thresholds_minor" : {
            "additionalProperties" : {
              "description" : "Action type to amount in minor units; below it no checker is needed (FR-APR-04)",
              "format" : "int64",
              "type" : "integer"
            },
            "description" : "Action type to amount in minor units; below it no checker is needed (FR-APR-04)",
            "type" : "object"
          },
          "approval_validity_days" : {
            "format" : "int32",
            "maximum" : 90,
            "minimum" : 1,
            "type" : "integer"
          },
          "disabled_collateral_types" : {
            "items" : {
              "pattern" : "^[a-z_]{2,40}$",
              "type" : "string"
            },
            "type" : "array"
          },
          "display_name" : {
            "maxLength" : 100,
            "minLength" : 1,
            "type" : "string"
          },
          "max_active_loans_per_member" : {
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          },
          "receipt_footer" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "require_mfa_all_staff" : {
            "type" : "boolean"
          },
          "retail_allow_negative_stock" : {
            "description" : "FR-RET-03: whether a retail sale may take stock below zero (default true)",
            "type" : "boolean"
          },
          "sms_sender_name" : {
            "description" : "Subject to aggregator approval",
            "pattern" : "^[A-Za-z0-9 ]{3,11}$",
            "type" : "string"
          },
          "sms_window_end" : {
            "pattern" : "^([01][0-9]|2[0-3]):[0-5][0-9]$",
            "type" : "string"
          },
          "sms_window_start" : {
            "pattern" : "^([01][0-9]|2[0-3]):[0-5][0-9]$",
            "type" : "string"
          }
        }
      },
...
    }
  },
  "info" : {
    "title" : "BMS Platform API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/retail/categories" : {
      "get" : {
        "operationId" : "listRetailCategories",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailCategoryList"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List product categories",
        "tags" : [ "retail-catalogue" ]
      },
      "post" : {
        "operationId" : "createRetailCategory",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RetailCategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailCategory"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Create a product category (FR-RET-01)",
        "tags" : [ "retail-catalogue" ]
      }
    },
...
    "/api/v1/retail/products" : {
      "get" : {
        "operationId" : "listRetailProducts",
        "parameters" : [ {
          "in" : "query",
          "name" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "category_id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "active",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "branch_id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailProductPage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Search products by code or description; with branch_id each row has that branch's balance",
        "tags" : [ "retail-catalogue" ]
      },
      "post" : {
        "operationId" : "createRetailProduct",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRetailProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailProduct"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Create a product with its first prices (FR-RET-01)",
        "tags" : [ "retail-catalogue" ]
      }
    },
    "/api/v1/retail/products/{product_id}" : {
      "get" : {
        "operationId" : "getRetailProduct",
        "parameters" : [ {
          "in" : "path",
          "name" : "product_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailProduct"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "A product",
        "tags" : [ "retail-catalogue" ]
      },
      "patch" : {
        "operationId" : "updateRetailProduct",
        "parameters" : [ {
          "in" : "path",
          "name" : "product_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateRetailProductRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailProduct"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edit a product's non-price fields",
        "tags" : [ "retail-catalogue" ]
      }
    },
    "/api/v1/retail/products/{product_id}/price-history" : {
      "get" : {
        "operationId" : "getRetailPriceHistory",
        "parameters" : [ {
          "in" : "path",
          "name" : "product_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailPriceHistory"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Every price change of a product (FR-RET-02)",
        "tags" : [ "retail-catalogue" ]
      }
    },
    "/api/v1/retail/products/{product_id}/prices" : {
      "post" : {
        "operationId" : "editRetailPrices",
        "parameters" : [ {
          "in" : "path",
          "name" : "product_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RetailPriceEditRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailProduct"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Edit prices by hand, with a history row (FR-RET-02)",
        "tags" : [ "retail-catalogue" ]
      }
    },
...
    "/api/v1/retail/units" : {
      "get" : {
        "operationId" : "listRetailUnits",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailUnitList"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List units of measure",
        "tags" : [ "retail-catalogue" ]
      },
      "post" : {
        "operationId" : "createRetailUnit",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RetailUnitRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RetailUnit"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Create a unit of measure (FR-RET-01)",
        "tags" : [ "retail-catalogue" ]
      }
    },
...
    "/api/v1/settings" : {
      "get" : {
        "operationId" : "getSettings",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantSettings"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Tenant settings",
        "tags" : [ "settings" ]
      },
      "patch" : {
        "operationId" : "updateSettings",
        "parameters" : [ {
          "in" : "header",
          "name" : "If-Match",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTenantSettingsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TenantSettings"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Change tenant settings, audited with before and after (FR-TEN-08)",
        "tags" : [ "settings" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}